chore: remove dead code

+++ src/users/users.service.ts
@@ -1,6 +1,5 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
-import { UpdateUserDto } from './dto/update-user.dto';
 import { Repository } from 'typeorm';
 import { User } from './entities/user.entity';
 import { InjectRepository } from '@nestjs/typeorm';
@@ -15,19 +14,28 @@ export class UsersService {
     return this.userRepository.save(createUserDto);
   }
 
-  findAll() {
-    return this.userRepository.find();
+  async findByEmail(email: string): Promise<User> {
+    const customer = await this.userRepository.findOneBy({ email });
+
+    if (!customer) {
+      throw new NotFoundException('Customer with this email does not exist');
+    }
+
+    return customer;
   }
 
-  findOne(email: string) {
-    return this.userRepository.findOneBy({ email });
-  }
-
-  update(id: number, updateUserDto: UpdateUserDto) {
-    return this.userRepository.update(id, updateUserDto);
+  findOne(id: number) {
+    const user = this.userRepository.findOneBy({ id });
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+    return user;
   }
 
   remove(id: number) {
+    if (!this.userRepository.findOneBy({ id })) {
+      throw new NotFoundException('User not found');
+    }
     return this.userRepository.delete(id);
   }
 }
